Add files via upload
Get ready to watch, remember, repeat! The Simon game is the exciting electronic game of lights and sounds in which players must repeat random sequences of lights by pressing the colored pads in the correct order. It's fast-paced play, with lights and sounds that can challenge you. Experience the fun as you repeat the patterns and advance to higher levels. Keep track of your score as you challenge friends or try to beat your own high score. Includes Simon game unit and instructions. 

Classic Simon gameplay 
Follow lights and sounds 
Suspense builds as sequences get longer 
Play solo or challenge a friend 
For ages 8 and up